src/backend/pyproject.toml
[tool.commitizen]
version = "0.0.17"
version_files = [
    "pyproject.toml:version"
]
update_changelog_on_bump = true
annotated_tag = true

[tool.poetry]
name = "asceticism"
version = "0.0.17"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.0.3"
psycopg2-binary = "^2.9.6"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
alembic = "^1.10.3"
sqlalchemy = "^2.0.9"
passlib = "^1.7.4"
pyjwt = "^2.6.0"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
httpx = "^0.24.0"


[tool.poetry.group.dev.dependencies]
commitizen = "^2.42.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
